chore: remove leftover state from inventory page
