+++ website/app/page.tsx
@@ -1,415 +1,71 @@
 "use client"
 
-import { useState, useEffect } from "react"
+import { useState } from "react"
 import { Card, CardContent } from "@/components/ui/card"
 import { Input } from "@/components/ui/input"
 import { Search, BarChart2, CalendarIcon, Github, X, TrendingUp, Clock, Package } from "lucide-react"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { motion, AnimatePresence } from "framer-motion"
+import { colors, categoryColors } from "@/constants/colors"
+import { rankingColors } from "@/constants/rankings"
+import {
+  categorizeItems,
+  getCategoryTotals,
+  generatePieChart,
+  getItemColor,
+  getTotalItemCount,
+  getUniqueItemCount,
+  getTopItems,
+  filterAndSortItems,
+} from "@/utils/inventory"
+import { useDateList } from "@/hooks/use-date-list"
+import { useInventoryData } from "@/hooks/use-inventory-data"
+import { useItemTrend } from "@/hooks/use-item-trend"
+import { useStatTrend } from "@/hooks/use-stat-trend"
 
-interface InventoryData {
-  item_count: Record<string, number>
-  duration: string
-}
+  export default function InventoryPage() {
+    const {
+      dateList,
+      selectedDate,
+      setSelectedDate,
+      handleSelectAll,
+      loading: dateListLoading,
+      error: dateListError,
+    } = useDateList()
+    const {
+      data,
+      loading: dataLoading,
+      error: dataError,
+    } = useInventoryData(selectedDate)
+    const {
+      selectedItem,
+      itemTrendData,
+      loadingTrend,
+      handleItemClick,
+      closeItemModal,
+    } = useItemTrend()
+    const { statModal, handleStatClick, closeStatModal } = useStatTrend()
+    const [searchTerm, setSearchTerm] = useState("")
+    const [isChangingDate, setIsChangingDate] = useState(false)
+    const [animationKey, setAnimationKey] = useState(0)
+    const [hoverIndex, setHoverIndex] = useState<number | null>(null)
+    const loading = dateListLoading || dataLoading
+    const error = dateListError || dataError
 
-interface DateItem {
-  value: string
-  label: string
-}
-
-interface CategoryTotal {
-  name: string
-  count: number
-  color: string
-}
-
-interface ItemTrendData {
-  [date: string]: number
-}
-
-interface StatModalData {
-  type: "totalItems" | "duration" | "item"
-  title: string
-  data: ItemTrendData | null
-  loading: boolean
-}
-
-export default function InventoryPage() {
-  const [dateList, setDateList] = useState<DateItem[]>([])
-  const [selectedDate, setSelectedDate] = useState<string>("") 
-  const [data, setData] = useState<InventoryData | null>(null)
-  const [loading, setLoading] = useState(true)
-  const [error, setError] = useState<string | null>(null)
-  const [searchTerm, setSearchTerm] = useState("")
-  const [isChangingDate, setIsChangingDate] = useState(false)
-  const [animationKey, setAnimationKey] = useState(0)
-  const [hoverIndex, setHoverIndex] = useState<number | null>(null)
-  const [selectedItem, setSelectedItem] = useState<string | null>(null)
-  const [itemTrendData, setItemTrendData] = useState<ItemTrendData | null>(null)
-  const [loadingTrend, setLoadingTrend] = useState(false)
-  const [statModal, setStatModal] = useState<StatModalData | null>(null)
-
-
-  // 定义颜色方案
-  const colors = {
-    primary: "#3498db", // 主色调：蓝色
-    secondary: "#2c3e50", // 次要色调：深蓝
-    accent1: "#4CAF50", // 强调色1：绿色
-    accent2: "#f39c12", // 强调色2：橙色
-    light: "#e6f2ff", // 浅色背景
-    lightBorder: "#85b8e8", // 浅色边框
-    lightGray: "#f5f5f5", // 浅灰色
-    mediumGray: "#bdc3c7", // 中灰色
-    darkText: "#003366", // 深色文字
-  }
-
-  // 定义类别颜色
-  const categoryColors = {
-    圣遗物: "#34495e", // 深蓝
-    矿物: "#4CAF50", // 绿色
-    食材: "#f39c12", // 橙色
-    其他: "#bdc3c7", // 浅灰
-  }
-
-  // 定义排行榜颜色
-  const rankingColors = [
-    "#3498db", // 蓝色
-    "#27ae60", // 绿色
-    "#f39c12", // 橙色
-    "#9b59b6", // 紫色
-    "#e74c3c", // 红色
-    "#1abc9c", // 青色
-    "#d35400", // 深橙色
-    "#2980b9", // 深蓝色
-    "#8e44ad", // 深紫色
-    "#c0392b", // 深红色
-  ]
-
-  // 处理统计卡片点击
-  const handleStatClick = async (type: "totalItems" | "duration") => {
-    const titles = {
-      totalItems: "总物品数量趋势",
-      duration: "运行时间趋势",
-    }
-
-    setStatModal({
-      type,
-      title: titles[type],
-      data: null,
-      loading: true,
-    })
-
-    try {
-      // 调用相应的API获取趋势数据
-      const apiEndpoint = type === "totalItems" ? "/api/total-items-trend" : "/api/duration-trend"
-      const response = await fetch(apiEndpoint)
-
-      if (response.ok) {
-        const trendData = await response.json()
-        setStatModal((prev) => (prev ? { ...prev, data: trendData.data, loading: false } : null))
-      } else {
-        // 如果API调用失败，使用模拟数据
-        const mockTrendData: ItemTrendData = generateMockTrendData(type)
-        setStatModal((prev) => (prev ? { ...prev, data: mockTrendData, loading: false } : null))
-      }
-    } catch (error) {
-      console.error(`获取${titles[type]}数据失败:`, error)
-      // 使用模拟数据作为备选
-      const mockTrendData: ItemTrendData = generateMockTrendData(type)
-      setStatModal((prev) => (prev ? { ...prev, data: mockTrendData, loading: false } : null))
-    }
-  }
-
-  // 生成模拟趋势数据
-  const generateMockTrendData = (type: "totalItems" | "duration"): ItemTrendData => {
-    const dates = [
-      "20250520",
-    ]
-    const mockData: ItemTrendData = {"20250520":10}
-    return mockData
-  }
-
-  // 处理物品点击
-  const handleItemClick = async (itemName: string) => {
-    setSelectedItem(itemName)
-    setLoadingTrend(true)
-
-    try {
-      // 调用API获取物品趋势数据
-      const response = await fetch(`/api/item-trend?item=${encodeURIComponent(itemName)}`)
-      if (response.ok) {
-        const trendData = await response.json()
-        setItemTrendData(trendData.data)
-      } else {
-        // 如果API调用失败，使用模拟数据
-        const mockTrendData: ItemTrendData = {
-          "2025-05-20": 11,
-        }
-        setItemTrendData(mockTrendData)
-      }
-    } catch (error) {
-      console.error("获取物品趋势数据失败:", error)
-      // 使用模拟数据作为备选
-      const mockTrendData: ItemTrendData = {
-        "2025-05-20": 11,
-      }
-      setItemTrendData(mockTrendData)
-    } finally {
-      setLoadingTrend(false)
-    }
-  }
-
-  // 关闭物品详情模态框
-  const closeItemModal = () => {
-    setSelectedItem(null)
-    setItemTrendData(null)
-  }
-    // 关闭统计模态框
-  const closeStatModal = () => {
-    setStatModal(null)
-  }
-
-    // 处理日期变化
   const handleDateChange = (newDate: string) => {
-      if (newDate !== selectedDate) {
-        setIsChangingDate(true)
-  
-        // 短暂延迟以允许退出动画完成
-        setTimeout(() => {
-          setSelectedDate(newDate)
-          setAnimationKey((prev) => prev + 1)
-  
-          // 模拟数据加载
-          setLoading(true)
-          setTimeout(() => {
-            setLoading(false)
-            setIsChangingDate(false)
-          }, 600)
-        }, 300)
-      }
-    }
-
-  // 从API获取日期列表
-  useEffect(() => {
-    const fetchDateList = async () => {
-      try {
-        setLoading(true)
-        const response = await fetch('/api/LogList')
-        
-        if (!response.ok) {
-          throw new Error(`HTTP error! status: ${response.status}`)
-        }
-        
-        const responseData = await response.json()
-        // 从返回的JSON中提取日期列表
-        const dates = responseData.list || []
-        // 将日期转换为Select组件需要的格式
-        const formattedDates = Array.isArray(dates) ? dates.map(date => ({
-          value: date,
-          label: date
-        })) : []
-        
-        setDateList(formattedDates)
-        
-        // 设置第一个日期为默认选中日期
-        if (formattedDates.length > 0) {
-          setSelectedDate(formattedDates[0].value)
-        }
-        
-        setError(null)
-      } catch (error) {
-        console.error('Error fetching date list:', error)
-        setError('获取日期列表失败，请稍后再试')
-      } finally {
-        setLoading(false)
-      }
-    }
-
-    fetchDateList()
-  }, [])
-
-  const handleSelectAll = () => {
-    setSelectedDate('all')  // 设定一个特殊值来表示“全部”
-  }
-
-  // 获取物品数据
-  useEffect(() => {
-    const fetchData = async () => {
-      if (!selectedDate) return
-      
-      try {
-        setLoading(true)
-        const response = await fetch(`/api/analyse?date=${selectedDate}`)
-        
-        if (!response.ok) {
-          throw new Error(`HTTP error! status: ${response.status}`)
-        }
-        
-        const data = await response.json()
-        setData(data)
-        setError(null)
-      } catch (error) {
-        console.error('Error fetching data:', error)
-        setError('获取数据失败，请稍后再试')
-      } finally {
-        setLoading(false)
-      }
-    }
-
-    fetchData()
-  }, [selectedDate])
-
-  // 物品分类逻辑
-  const getCategory = (name: string) => {
-    if (
-      name.includes("冒险家") ||
-      name.includes("游医") ||
-      name.includes("幸运儿") ||
-      name.includes("险家") ||
-      name.includes("医的") ||
-      name.includes("运儿") ||
-      name.includes("家") ||
-      (name.includes("的") &&
-        (name.includes("方巾") ||
-          name.includes("枭羽") ||
-          name.includes("怀钟") ||
-          name.includes("药壶") ||
-          name.includes("银莲") ||
-          name.includes("怀表") ||
-          name.includes("尾羽") ||
-          name.includes("头带") ||
-          name.includes("金杯") ||
-          name.includes("之花") ||
-          name.includes("之杯") ||
-          name.includes("沙漏") ||
-          name.includes("绿花") ||
-          name.includes("银冠") ||
-          name.includes("鹰羽")))
-    ) {
-      return "圣遗物"
-    } else if (['铁块', '白铁块', '水晶块', '魔晶块', '星银矿石', '紫晶块', '萃凝晶'].includes(name)) {
-      return "矿物"
-    } else if (
-      ['苹果', '蘑菇', '甜甜花', '胡萝卜', '白萝卜', '金鱼草', '薄荷',
-         '松果', '树莓', '松茸', '鸟蛋', '海草', '堇瓜', '墩墩桃',
-          '须弥蔷薇', '枣椰', '茉洁草', '沉玉仙茗', '颗粒果', '澄晶实',
-           '红果果菇', '小灯草', '嘟嘟莲', '莲蓬', '绝云椒椒', '清心',
-            '马尾', '琉璃袋', '竹笋', '绯樱绣球', '树王圣体菇', '帕蒂沙兰',
-             '青蜜莓'].includes(name)
-    ) {
-      return "食材"
-    } else {
-      return "其他"
+    if (newDate !== selectedDate) {
+      setIsChangingDate(true)
+      setTimeout(() => {
+        setSelectedDate(newDate)
+        setAnimationKey((prev) => prev + 1)
+        setIsChangingDate(false)
+      }, 300)
     }
   }
 
-  // 对物品进行分类
-  const categorizeItems = (items: Record<string, number>) => {
-    const categories: Record<string, Record<string, number>> = {
-      圣遗物: {},
-      矿物: {},
-      食材: {},
-      其他: {},
-    }
-
-    Object.entries(items).forEach(([name, count]) => {
-      let category = getCategory(name)
-      categories[category][name] = count
-    })
-
-    return categories
-  }
-
-    // 计算每个分类的总数量
-    const getCategoryTotals = (categories: Record<string, Record<string, number>>) => {
-  
-      const totals: CategoryTotal[] = []
-  
-      Object.entries(categories).forEach(([name, items]) => {
-        const count = Object.values(items).reduce((sum, count) => sum + count, 0)
-        totals.push({
-          name,
-          count,
-          color: categoryColors[name as keyof typeof categoryColors],
-        })
-      })
-  
-      return totals
-    }
-
-  // 过滤和排序物品
-  const getTop5Items = (items: Record<string, number>) => {
-    return Object.entries(items)
-      .sort((a, b) => b[1] - a[1])
-      .slice(0, 15)
-  }
-
-  // 过滤和排序物品 (默认按数量排序)
-  const filterAndSortItems = (items: Record<string, number>) => {
-    return Object.entries(items)
-      .filter(([name]) => name.toLowerCase().includes(searchTerm.toLowerCase()))
-      .sort((a, b) => b[1] - a[1])
-  }
-
-  // 获取总物品数量
-  const getTotalItemCount = (items: Record<string, number>) => {
-    return Object.values(items).reduce((sum, count) => sum + count, 0)
-  }
-
-  // 获取物品种类数量
-  const getUniqueItemCount = (items: Record<string, number>) => {
-    return Object.keys(items).length
-  }
-
-  const generatePieChart = (data: CategoryTotal[]) => {
-    const total = data.reduce((sum, item) => sum + item.count, 0)
-
-    // 检查是否只有一个非 0 数据项
-    const nonZeroItems = data.filter(item => item.count > 0)
-    if (nonZeroItems.length === 1) {
-      const item = nonZeroItems[0]
-      return [{
-        isFullCircle: true,
-        color: item.color,
-        name: item.name,
-        count: item.count,
-        percentage: '100.0',
-        path: "", // 添加空路径属性以满足类型要求
-      }]
-    }
-
-    let currentAngle = 0
-
-    return data.map((item) => {
-      const percentage = item.count / total
-      const startAngle = currentAngle
-      const endAngle = currentAngle + percentage * 2 * Math.PI
-
-      const x1 = 100 + 80 * Math.cos(startAngle)
-      const y1 = 100 + 80 * Math.sin(startAngle)
-      const x2 = 100 + 80 * Math.cos(endAngle)
-      const y2 = 100 + 80 * Math.sin(endAngle)
-
-      const largeArcFlag = percentage > 0.5 ? 1 : 0
-
-      const path = `M 100 100 L ${x1} ${y1} A 80 80 0 ${largeArcFlag} 1 ${x2} ${y2} Z`
-
-      currentAngle = endAngle
-
-      return {
-        path,
-        color: item.color,
-        name: item.name,
-        count: item.count,
-        percentage: (percentage * 100).toFixed(1),
-      }
-    })
-  }
-
-  const getItemColor = (name: string) => {
-    let category = getCategory(name)
-    return categoryColors[category as keyof typeof categoryColors]
+  const handleSegmentHover = (index: number | null) => {
+    setHoverIndex(index)
   }
  // 动画变体
  const containerVariants = {
@@ -496,14 +152,9 @@ const chartVariants = {
   const totalItems = getTotalItemCount(data.item_count)
   const uniqueItems = getUniqueItemCount(data.item_count)
   const duration = data.duration
-  const top10Items = getTop5Items(data.item_count)
+  const top10Items = getTopItems(data.item_count)
   const categoryTotals = getCategoryTotals(categories)
   const pieChartData = generatePieChart(categoryTotals)
-
-   // 新增：处理扇形区域悬停
-  const handleSegmentHover = (index: number | null) => {
-    setHoverIndex(index)
-  }
 
   // 找出最大数量，用于计算条形图比例
   const maxCount = top10Items[0][1]
