+++ website/constants/colors.ts
@@ -1,0 +1,18 @@
+export const colors = {
+  primary: "#3498db", // 主色调：蓝色
+  secondary: "#2c3e50", // 次要色调：深蓝
+  accent1: "#4CAF50", // 强调色1：绿色
+  accent2: "#f39c12", // 强调色2：橙色
+  light: "#e6f2ff", // 浅色背景
+  lightBorder: "#85b8e8", // 浅色边框
+  lightGray: "#f5f5f5", // 浅灰色
+  mediumGray: "#bdc3c7", // 中灰色
+  darkText: "#003366", // 深色文字
+}
+
+export const categoryColors = {
+  圣遗物: "#34495e", // 深蓝
+  矿物: "#4CAF50", // 绿色
+  食材: "#f39c12", // 橙色
+  其他: "#bdc3c7", // 浅灰
+}
