+++ website/constants/rankings.ts
@@ -1,0 +1,12 @@
+export const rankingColors = [
+  "#3498db", // 蓝色
+  "#27ae60", // 绿色
+  "#f39c12", // 橙色
+  "#9b59b6", // 紫色
+  "#e74c3c", // 红色
+  "#1abc9c", // 青色
+  "#d35400", // 深橙色
+  "#2980b9", // 深蓝色
+  "#8e44ad", // 深紫色
+  "#c0392b", // 深红色
+]
