+++ website/hooks/use-item-trend.ts
@@ -1,0 +1,41 @@
+import { useState } from "react"
+import { ItemTrendData } from "@/types/inventory"
+
+export function useItemTrend() {
+  const [selectedItem, setSelectedItem] = useState<string | null>(null)
+  const [itemTrendData, setItemTrendData] = useState<ItemTrendData | null>(null)
+  const [loadingTrend, setLoadingTrend] = useState(false)
+
+  const handleItemClick = async (itemName: string) => {
+    setSelectedItem(itemName)
+    setLoadingTrend(true)
+
+    try {
+      const response = await fetch(`/api/item-trend?item=${encodeURIComponent(itemName)}`)
+      if (response.ok) {
+        const trendData = await response.json()
+        setItemTrendData(trendData.data)
+      } else {
+        const mockTrendData: ItemTrendData = {
+          "2025-05-20": 11,
+        }
+        setItemTrendData(mockTrendData)
+      }
+    } catch (error) {
+      console.error("获取物品趋势数据失败:", error)
+      const mockTrendData: ItemTrendData = {
+        "2025-05-20": 11,
+      }
+      setItemTrendData(mockTrendData)
+    } finally {
+      setLoadingTrend(false)
+    }
+  }
+
+  const closeItemModal = () => {
+    setSelectedItem(null)
+    setItemTrendData(null)
+  }
+
+  return { selectedItem, itemTrendData, loadingTrend, handleItemClick, closeItemModal }
+}
